Mod updated for 1.4 and passed autotests. LanguageData LanguageData LanguageData LanguageData Used DeepL to update translations for ChineseSimplified, French, German, Russian Remaining characters: 487821
