+++ About/Changelog.txt
@@ -1,4 +1,9 @@
 # Changelog for TrulyInspiredTaming
 
+1.4.1 - Mod updated for 1.4 and passed autotests.
+LanguageData LanguageData LanguageData LanguageData Used DeepL to update translations for ChineseSimplified, French, German, Russian 
+Remaining characters: 487821
+
+
 1.3.1 - Updated to 1.3
