+++ About/Manifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <Manifest>
   <identifier>TrulyInspiredTaming</identifier>
-  <version>1.3.1</version>
+  <version>1.4.1</version>
   <dependencies />
   <incompatibleWith />
   <loadBefore />
