+++ About/ModSync.xml
@@ -2,7 +2,7 @@
 <ModSyncNinjaData>
   <ID>6f3ffbcd-d62f-43e5-9d68-46a351909d7d</ID>
   <ModName>Truly Inspired Taming (Continued)</ModName>
-  <Version>1.3.1</Version>
+  <Version>1.4.1</Version>
   <SaveBreaking>False</SaveBreaking>
   <Host name="Github">
     <Owner>emipa606</Owner>
@@ -12,8 +12,3 @@
     <Branch>master</Branch>
   </Host>
 </ModSyncNinjaData>
-
-
-
-
-
