+++ Languages/ChineseSimplified/Keyed/Keys.xml
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<LanguageData>
+  <!--The following translations were generated by https://www.deepl.com/-->
+  <TIT_CurrentModVersion>已安装的mod-version。{0}</TIT_CurrentModVersion>
+  <TIT_Title>真正有灵感的驯服</TIT_Title>
+  <TIT_TamingSkillBoostType>鼓舞人心的驯服等级提升</TIT_TamingSkillBoostType>
+  <TIT_SettingsExplanation>拥有 "灵感驯服 "的殖民者现在可以驯服超出他们当前技能水平的动物。  这些选项决定了超越的程度。</TIT_SettingsExplanation>
+  <TIT_TameAnyAnimal>驯服任何动物</TIT_TameAnyAnimal>
+  <TIT_BoostByPercentage>按百分比提升</TIT_BoostByPercentage>
+  <TIT_BoostByLevels>按级别提升</TIT_BoostByLevels>
+  <TIT_AnimalSkillStillTooLow>动物技能太低--需要{0}（有{1}{2}的启发提升）。</TIT_AnimalSkillStillTooLow>
+</LanguageData>
