+++ Languages/French/Keyed/Keys.xml
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<LanguageData>
+  <!--The following translations were generated by https://www.deepl.com/-->
+  <TIT_CurrentModVersion>Version installée du module : {0}</TIT_CurrentModVersion>
+  <TIT_Title>Une domestication vraiment inspirée</TIT_Title>
+  <TIT_TamingSkillBoostType>Augmentation du niveau d'apprivoisement inspiré</TIT_TamingSkillBoostType>
+  <TIT_SettingsExplanation>Les colons ayant reçu une formation inspirée peuvent désormais apprivoiser des animaux au-delà de leur niveau de compétence actuel.  Ces options déterminent jusqu'à quel point.</TIT_SettingsExplanation>
+  <TIT_TameAnyAnimal>Apprivoiser n'importe quel animal</TIT_TameAnyAnimal>
+  <TIT_BoostByPercentage>Augmentation en pourcentage</TIT_BoostByPercentage>
+  <TIT_BoostByLevels>Booster par niveaux</TIT_BoostByLevels>
+  <TIT_AnimalSkillStillTooLow>compétence animaux trop faible - besoin de {0} (a {1} {2} boost inspiré)</TIT_AnimalSkillStillTooLow>
+</LanguageData>
