+++ Languages/German/Keyed/Keys.xml
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<LanguageData>
+  <!--The following translations were generated by https://www.deepl.com/-->
+  <TIT_CurrentModVersion>Installierte Mod-Version: {0}</TIT_CurrentModVersion>
+  <TIT_Title>Wahrhaftig inspirierte Zähmung</TIT_Title>
+  <TIT_TamingSkillBoostType>Inspirierte Zähmung: Stufenanstieg</TIT_TamingSkillBoostType>
+  <TIT_SettingsExplanation>Kolonisten mit "Inspiriertes Zähmen" können jetzt Tiere zähmen, die über ihre aktuelle Fertigkeitsstufe hinausgehen.  Diese Optionen bestimmen, wie weit darüber hinaus.</TIT_SettingsExplanation>
+  <TIT_TameAnyAnimal>Jedes Tier zähmen</TIT_TameAnyAnimal>
+  <TIT_BoostByPercentage>Steigerung in Prozent</TIT_BoostByPercentage>
+  <TIT_BoostByLevels>Erhöhung um Stufen</TIT_BoostByLevels>
+  <TIT_AnimalSkillStillTooLow>Tierfertigkeit zu niedrig - braucht {0} (hat {1} {2} inspirierte Verstärkung)</TIT_AnimalSkillStillTooLow>
+</LanguageData>
