+++ Languages/Russian/Keyed/Keys.xml
@@ -1,0 +1,12 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<LanguageData>
+  <!--The following translations were generated by https://www.deepl.com/-->
+  <TIT_CurrentModVersion>Установленная версия мода: {0}</TIT_CurrentModVersion>
+  <TIT_Title>Поистине вдохновенное укрощение</TIT_Title>
+  <TIT_TamingSkillBoostType>Повышение уровня воодушевленного укрощения</TIT_TamingSkillBoostType>
+  <TIT_SettingsExplanation>Колонисты с Вдохновенным приручением теперь могут приручать животных, превосходящих их текущий уровень навыка.  Эти параметры определяют, насколько выше.</TIT_SettingsExplanation>
+  <TIT_TameAnyAnimal>Приручить любое животное</TIT_TameAnyAnimal>
+  <TIT_BoostByPercentage>Увеличение в процентах</TIT_BoostByPercentage>
+  <TIT_BoostByLevels>Повышение на уровни</TIT_BoostByLevels>
+  <TIT_AnimalSkillStillTooLow>навык животных слишком низок - нужно {0} (есть {1} {2} вдохновляющее усиление)</TIT_AnimalSkillStillTooLow>
+</LanguageData>
